refactor: fold CLI agent resolution into agent catalog
