+++ packages/cli/src/commands/agents.ts
@@ -7,6 +7,7 @@ import {
   formatCliHiddenAgentsNotice,
   loadCliAgentList,
   parseCliAgentCategoryFilter,
+  resolveCliAgent,
   type CliAgentListOptions,
 } from '../runtime/agents';
 import { CliExitCode } from '../runtime/exitCodes';
@@ -20,7 +21,6 @@ import {
 import { defineCliCommand } from './_helpers/defineCliCommand';
 import { GLOBAL_ARGS, optString } from './_helpers/globalArgs';
 import { emitCliResult } from './_helpers/output';
-import { resolveCliAgent } from '../runtime/agentResolution';
 import { agentsRunCommand } from './agentsRun';
 import type { CliContext } from '../runtime/cliContext';
