+++ src/test-kernel/cli/AgentsRunCommand.vitest.mts
@@ -49,7 +49,7 @@ vi.mock('@cli/commands/_helpers/output', () => ({
   emitCliResult: vi.fn(),
 }));
 
-vi.mock('@cli/runtime/agentResolution', () => ({
+vi.mock('@cli/runtime/agents', () => ({
   resolveCliAgent: mocks.resolveCliAgent,
 }));
