+++ src/test-kernel/cli/WorkflowRunCommand.vitest.mts
@@ -46,7 +46,7 @@ vi.mock('@cli/commands/_helpers/output', () => ({
   emitCliResult: vi.fn(),
 }));
 
-vi.mock('@cli/runtime/agentResolution', () => ({
+vi.mock('@cli/runtime/agents', () => ({
   resolveCliAgent: mocks.resolveCliAgent,
 }));
